Reorganize auction and adunit components
