+++ packages/extension/src/view/devtools/components/privateAdvertising/protectedAudience/auctions/container.tsx
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2024 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * External dependencies.
+ */
+import { SidebarProvider, type SidebarItems } from '@google-psat/design-system';
+import React from 'react';
+
+/**
+ * Internal dependencies.
+ */
+import AuctionPanel from './panel';
+import type { AuctionEventsType } from '../../../../stateProviders/protectedAudience/context';
+
+interface AuctionsContainerProps {
+  auctionEvents: AuctionEventsType;
+  sidebarData: SidebarItems;
+  setSidebarData: React.Dispatch<React.SetStateAction<SidebarItems>>;
+}
+
+const AuctionsContainer = ({
+  auctionEvents,
+  sidebarData,
+  setSidebarData,
+}: AuctionsContainerProps) => {
+  return (
+    <SidebarProvider
+      data={sidebarData}
+      defaultSelectedItemKey={Object.keys(sidebarData)?.[0]}
+    >
+      <AuctionPanel
+        setSidebarData={setSidebarData}
+        auctionEvents={auctionEvents}
+      />
+    </SidebarProvider>
+  );
+};
+
+export default AuctionsContainer;
