+++ packages/extension/src/view/devtools/components/privateAdvertising/protectedAudience/auctions/table/bottomTray.tsx
